% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_os.R
\name{calcule_surface}
\alias{calcule_surface}
\title{Calcule the Surface}
\usage{
calcule_surface(path_pdb, online)
}
\arguments{
\item{path_pdb}{Path or name of PDB File}

\item{online}{Field to define if the file will be otain in a online way.}
}
\description{
This function, implements and execute all functions essentials
to run the OSFIBO code. One can reorganize the PDB file, to generate datas
about contacts of proteins and theirs areas of contact. The return of function is
a group of informations that contains all areas and contacts.
}
\author{
Herson Hebert
}
