% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaner.R
\name{clean_pdb}
\alias{clean_pdb}
\title{PDB Cleaning.}
\usage{
clean_pdb(path_pdb, online)
}
\arguments{
\item{path_pdb}{Name or path of PDB file.}

\item{online}{Logical parameter that defines whether to download or locally load the PDB file.}
}
\description{
Cleaning PDB performs the reorganization of data from an input
PDB file. This function is responsible for removing lines and
information that are not necessary for calculating occluded
areas. In addition, the residues and their atoms are renumbered
to perform the area calculations.
}
\author{
Carlos Henrique da Silveira

Herson Hebert Mendes Soares

João Paulo Roquim Romanelli
}
