% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folders_manipulate.R
\name{create_folder}
\alias{create_folder}
\title{Files Manipulating.}
\usage{
create_folder()
}
\description{
Function for creating folders and manipulating files in the FIBOS system.
}
\author{
Carlos Henrique da Silveira

Herson Hebert Mendes Soares

João Paulo Roquim Romanelli

Patrick Fleming
}
