% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_intermediary.R
\name{execute_os}
\alias{execute_os}
\alias{call_main}
\title{Reading Values of Occluded Surface Area.}
\usage{
call_main(iresf, iresl, maxres, maxat)
}
\arguments{
\item{iresf}{Name or path of prot.srf.}

\item{iresl}{Name or path of prot.srf.}

\item{maxres}{Maximum number of residues.}

\item{maxat}{Maximum number of atoms.}
}
\description{
It's a important function to load calculated values to use of user.
}
\author{
Herson Hebert
}
