% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_os.R
\name{occluded_surface}
\alias{occluded_surface}
\title{Calcule the Surface}
\usage{
occluded_surface(path_pdb, online, method)
}
\arguments{
\item{path_pdb}{Path or name of PDB File}

\item{online}{Logical parameter that defines whether to download or locally load the PDB file.}

\item{method}{Method to be used - 1: OS ; 2: FIBOS}
}
\description{
The calculation of occluded surface areas is essential for
understanding the possibility of an enzyme passing between atoms
of a protein. To perform the calculation, it is considered that
a surface is occluded based on tests with a probe, which is
typically the water molecule.
}
\seealso{
\code{\link[=read_prot]{read_prot()}}
}
\author{
Carlos Henrique da Silveira

Herson Hebert Mendes Soares

João Paulo Roquim Romanelli
}
