% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaner.R
\name{clean_pdb}
\alias{clean_pdb}
\title{Reorganization of PDB file.}
\usage{
clean_pdb(path_pdb, online)
}
\arguments{
\item{path_pdb}{Name or path of PDB file.}

\item{online}{Tag to describes if the read will be obtained of Online Way.}
}
\description{
It's a necessary function to help in the proccess to read file and calculate areas.
}
\author{
Herson Hebert
}
