% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/respak.R
\name{respak}
\alias{respak}
\title{Respak Calcule}
\usage{
respak()
}
\arguments{
\item{prot}{Prot File.}
}
\description{
The implemented function executes the implemented methods.
Using this function, it is possible to calculate occluded areas
through the traditional methodology, Occluded Surface, or by
applying the Fibonacci OS methodology. At the end of the method
execution, the "prot.srf" file is generated, and returned for
the function. The data in this file refers to all contacts
between atoms of molecules present in a protein's PDB.
}
\seealso{
os::read_OS

os::occluded_surface
}
\author{
Carlos Henrique da Silveira

Herson Hebert Mendes Soares

João Paulo Roquim Romanelli
}
